Broke up data, event, and transition parsing code.
